// src/shared/dataprovider/api/model/cardRequestDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AgentBankType } from './agentBankType';
import { AgentRemittance } from './agentRemittance';

export interface CardRequestDTO { 
    AgentId?: number;
    Type?: AgentBankType;
    Status?: AgentRemittance;
    CreatedDate?: Date;
    UpdatedDate?: Date;
    IsMainCard?: boolean;
    BankName?: string;
    AccountName?: string;
    AccountNumber?: string;
    NickName?: string;
    BinanceWalletAddress?: string;
}